Move WalletController -> WalletContainer
